(tae): txn refactor 5
