+++ pkg/vm/engine/tae/tables/txnentries/compactblkcmd.go
@@ -1,4 +1,4 @@
-package txnnodes
+package txnentries
 
 import "io"
