+++ pkg/vm/engine/tae/tables/txnentries/types.go
@@ -1,4 +1,4 @@
-package txnnodes
+package txnentries
 
 import "github.com/matrixorigin/matrixone/pkg/vm/engine/tae/iface/txnif"
